Add database initialization script
This commit adds a init_db file which will be used to
initialize the database on setting up the project
remotely or locally
