+++ server/src/datasources/index.js
@@ -1,3 +1,17 @@
 import PostgresAPI from './postgres';
 
-export default PostgresAPI;
+const config = {
+  user: process.env.PGUSER,
+  database: process.env.PGDATABASE,
+  password: process.env.PGPASSWORD,
+  port: 5432,
+  max: 10, // max number of clients in the pool
+  idleTimeoutMillis: 30000,
+};
+
+const dataSources = {
+  PostgresAPI,
+  config
+};
+
+export default dataSources;
