+++ server/src/graphql/context/index.js
@@ -1,14 +1,6 @@
-import PostgresAPI from '../../datasources';
+import dataSources from '../../datasources';
 
-const config = {
-  user: process.env.PGUSER,
-  database: process.env.PGDATABASE,
-  password: process.env.PGPASSWORD,
-  port: 5432,
-  max: 10, // max number of clients in the pool
-  idleTimeoutMillis: 30000,
-};
-
+const { config, PostgresAPI } = dataSources;
 const context = {
   client: new PostgresAPI(config),
 };
