+++ server/src/graphql/resolvers/schema.js
@@ -2,7 +2,7 @@ export const Schema = {
   users: (_, __, { client }) => {
     return client.getAllUsers();
   },
-  user: (_, __, { client }) => {
-    return client.getUser(1)
+  user: (_, { id }, { client }) => {
+    return client.getUser(id)
   }
 };
